# backend/docs/swagger.yaml
definitions:
...
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/cache/containers:
    get:
      consumes:
      - application/json
      description: 从缓存获取所有容器信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器缓存
      tags:
      - Admin API - 缓存管理
  /api/admin/cache/refresh:
    post:
      consumes:
      - application/json
      description: 刷新容器缓存，无name参数刷新所有，有name参数刷新指定容器
      parameters:
      - description: 容器名称，不传则刷新所有
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 刷新失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 刷新缓存
      tags:
      - Admin API - 缓存管理
...
  /api/admin/dashboard:
    get:
      consumes:
      - application/json
      description: 获取系统概览数据，包括容器、用户、系统信息等
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取仪表盘数据
      tags:
      - Admin API - 仪表盘
...
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: 使用用户名、密码和验证码登录管理后台
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          properties:
            captcha:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误或验证码错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      summary: 管理员登录
      tags:
      - Admin API - 认证
...
swagger: "2.0"